requirements.txt
riotwatcher
pandas
streamlit
plotly
python-dotenv
